requirements.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
anyio==3.3.4
argcomplete==2.0.0
argon2-cffi==21.1.0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@04f4dc2ca1ed0acc9893bd1a3b526a7e02c4bb10
cloudpickle==2.2.1
colorcet==3.0.1
cycler==0.10.0
Cython==0.29.33
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fsspec==2021.10.1
ftfy==6.0.3
future==0.18.2
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.8.0
hyperopt==0.2.7
idna==3.3
imageio==2.9.0
importlib-metadata==4.13.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.2.0
json5==0.9.6
jsonpickle==3.0.1
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyterlab==3.2.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
kneed==0.8.1
libmr==0.1.9
lmdb==1.4.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
ml-pyxis @ https://github.com/vicolab/ml-pyxis/archive/master.zip
msgpack==1.0.4
multidict==5.2.0
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
ninja==1.10.2.4
nltk==3.7
nose==1.3.7
notebook==6.4.4
numpy==1.21.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.1
opencv-python==4.7.0.68
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
param==1.12.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
portalocker==2.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
ptyprocess==0.7.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.5.0
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.4.9
pytz==2021.3
pyvis==0.3.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
torch==1.8.2+cu111
torch-encoding @ git+https://github.com/zhanghang1989/PyTorch-Encoding/@331ecdd5306104614cb414b16fbcd9d1a8d40e1e
torchaudio==0.8.2
torchmetrics==0.5.1
torchvision==0.9.2+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
yarl==1.7.0
zipp==3.10.0
